# openapi.yaml
openapi: 3.1.0
info:
  title: Atlas REI MVP API
  version: 1.0.0
  summary: MVP API for the Atlas REI real estate investment operating system
  description: >
    REST API for workspace management, investor profiles, underwriting engine,
    committee evaluation, portfolio tracking, and documents.
servers:
  - url: /api/v1
    description: Dev Proxy Prefix
paths:
...
  /workspaces/{workspaceId}/profiles:
    get:
      summary: List investor profiles in workspace
      tags:
        - Investor Profiles
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvestorProfile"
    post:
      summary: Create a new investor profile
      tags:
        - Investor Profiles
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvestorProfileRequest"
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestorProfile"
  /workspaces/{workspaceId}/profiles/{profileId}:
    put:
      summary: Update an existing investor profile
      tags:
        - Investor Profiles
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: profileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvestorProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestorProfile"
...
  /workspaces/{workspaceId}/shortlists:
    get:
      summary: Get watchlists/shortlists for workspace
      tags:
        - Shortlists
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of shortlists
    post:
      summary: Create a new shortlist
      tags:
        - Shortlists
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShortlistRequest"
      responses:
        "201":
          description: Shortlist created
  /workspaces/{workspaceId}/shortlists/{shortlistId}/items:
    post:
      summary: Add an item to a shortlist
      tags:
        - Shortlists
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: shortlistId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddShortlistItemRequest"
      responses:
        "201":
          description: Item added
  /workspaces/{workspaceId}/shortlists/{shortlistId}/items/{itemId}:
    delete:
      summary: Remove an item from a shortlist
      tags:
        - Shortlists
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: shortlistId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Item removed
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InvestorProfile:
      type: object
      required:
        - id
        - workspaceId
        - name
        - baseCurrency
        - capitalAvailable
        - riskTolerance
        - investmentHorizonMonths
        - incomeVsGrowthPreference
        - financingPreference
        - targetCountries
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
        baseCurrency:
          type: string
          maxLength: 3
        capitalAvailable:
          type: number
        riskTolerance:
          type: string
          enum: [conservative, moderate, aggressive]
        investmentHorizonMonths:
          type: integer
        incomeVsGrowthPreference:
          type: string
          enum: [income, growth, balanced]
        financingPreference:
          type: string
          enum: [cash, mortgage, flexible]
        targetCountries:
          type: array
          items:
            type: string
            maxLength: 2
        constraints:
          type: object
        goals:
          type: object
    CreateInvestorProfileRequest:
      type: object
      required:
        - name
        - capitalAvailable
        - riskTolerance
        - investmentHorizonMonths
        - incomeVsGrowthPreference
        - financingPreference
        - targetCountries
      properties:
        name:
          type: string
        baseCurrency:
          type: string
          maxLength: 3
          default: AED
        capitalAvailable:
          type: number
        riskTolerance:
          type: string
          enum: [conservative, moderate, aggressive]
        investmentHorizonMonths:
          type: integer
        incomeVsGrowthPreference:
          type: string
          enum: [income, growth, balanced]
        financingPreference:
          type: string
          enum: [cash, mortgage, flexible]
        targetCountries:
          type: array
          items:
            type: string
            maxLength: 2
        constraints:
          type: object
        goals:
          type: object
...
    CreateShortlistRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        investorProfileId:
          type: string
          format: uuid
    AddShortlistItemRequest:
      type: object
      required:
        - entityType
        - entityId
      properties:
        entityType:
          type: string
          enum: [property, district]
        entityId:
          type: string
        rankPosition:
          type: integer
        notes:
          type: string
...